+++ sharding-jdbc-ddl-parser/src/main/java/io/shardingsphere/core/parsing/parser/antler/sql/ddl/AlterTableStatement.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2016-2018 shardingsphere.io.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * </p>
+ */
+
+package io.shardingsphere.core.parsing.parser.antler.sql.ddl;
+
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+import io.shardingsphere.core.parsing.parser.sql.ddl.DDLStatement;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString(callSuper = true)
+public class AlterTableStatement extends DDLStatement {
+    private final List<String> dropColumns = new ArrayList<>();
+
+    private final Map<String, ColumnDefinition> updateColumns = new LinkedHashMap<>();
+
+    private final List<ColumnDefinition> addColumns = new ArrayList<>();
+
+    private boolean dropPrimaryKey;
+
+    private String newTableName;
+
+}
