--- sharding-jdbc-ddl-parser/src/main/java/io/shardingsphere/core/parsing/parser/sql/ddl/ColumnDefinition.java
@@ -1,20 +1,0 @@
-package io.shardingsphere.core.parsing.parser.sql.ddl;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@AllArgsConstructor
-@Getter
-@Setter
-@ToString
-public class ColumnDefinition {
-	private String name;
-	
-	private String type;
-	
-	private Integer length;
-	
-	private boolean primaryKey;
-}
