--- sharding-jdbc-ddl-parser/src/main/java/io/shardingsphere/core/parsing/parser/sql/ddl/DDLType.java
@@ -1,11 +1,0 @@
-package io.shardingsphere.core.parsing.parser.sql.ddl;
-
-public enum DDLType {
-	CREATETABLE,
-	ALTERTABLE,
-	DROPTABLE,
-	TRUNCATETABLE,
-	CREATEINDEX,
-	ALTERINDEX,
-	DROPINDEX,
-}
